OSIS-education_group_pagination Add page in search view
